func trendsControl
