+++ src/js/customFunction/getGenreName.js
@@ -1,6 +1,7 @@
-import { genreIdArr } from '../common';
+import { genreIdArr } from '../controls/trendsControl';
 
 export function getGenreName(genre_ids) {
+  console.log('aaa', genreIdArr);
   let genreName = [];
   genre_ids.forEach(genre_id => {
     console.log('ku', genreIdArr);
